requirements.txt
# Core AI and LLM libraries
google-generativeai>=0.3.2
openai>=1.0.0

# Web framework
streamlit>=1.28.1
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Data processing
pandas>=2.1.3
pydantic>=2.5.0
python-multipart>=0.0.6

# Document processing
python-docx>=0.8.11
PyPDF2>=3.0.1

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# File handling
pathlib

# Development tools (optional)
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1
pytest>=7.4.3

# Additional utilities
typing-extensions>=4.8.0
asyncio
json5>=0.9.14
